Initial render done through render loop
This means the first render is not special, so the program won't quit if
it fails. But we now have a file existence check, so there shouldn't be
a reason for it to fail. In the very worst case scenario, we'd get a
blank screen, but that shouldn't happen in practice.