Minor GTK tweaks
Newer versions of GTK don't have auto-option wrapping, so we need to put
`Some` automatically. This is not *necessary* right now, but would be if
we updated.

Currently, doesn't seem to be possible to update, because webkit2gtk and
the other GTK stuff needs to use the same glib version and that doesn't
seem to be the case. Might be a solveable problem, but I don't really
need anything from the newer versions that I know of.